feat(miette): use miette for pretty errors
This commit:

- factors out all the custom error types into a new error module
- implements help diagnostics for said errors
- sets up the main app to make it ready to render the miette errors
- does some custom displaying of the miette errors in case the user
  chose the repl mode
- converts some std error types in the one shot mode for compatibility
  reasons

Authored-by: RobWalt <robwalter96@gmail.com>
