+++ Cargo.toml
@@ -11,6 +11,7 @@ anyhow = "1.0.79"
 ariadne = "0.4.0"
 chumsky = "0.9.3"
 lazy_static = "1.4.0"
+miette = { version = "7.0.0", features = ["fancy"] }
 ordered-float = "4.2.0"
 serde_json = "1.0.111"
 thiserror = "1.0.56"
